@@ -1,17 +1,24 @@
 module gotest.tools/gotestsum
 
+go 1.21
+
 require (
+	github.com/bitfield/gotestdox v0.2.2
 	github.com/dnephin/pflag v1.0.7
-	github.com/fatih/color v1.13.0
-	github.com/fsnotify/fsnotify v1.5.4
-	github.com/google/go-cmp v0.5.8
+	github.com/fatih/color v1.17.0
+	github.com/fsnotify/fsnotify v1.8.0
+	github.com/google/go-cmp v0.6.0
 	github.com/google/shlex v0.0.0-20191202100458-e7afc7fbc510
-	github.com/mattn/go-colorable v0.1.12 // indirect
-	golang.org/x/sync v0.0.0-20220601150217-0de741cfad7f
-	golang.org/x/sys v0.0.0-20220715151400-c0bba94af5f8
-	golang.org/x/term v0.0.0-20220526004731-065cf7ba2467
-	golang.org/x/tools v0.1.11
-	gotest.tools/v3 v3.3.0
+	golang.org/x/sync v0.8.0
+	golang.org/x/sys v0.30.0
+	golang.org/x/term v0.29.0
+	golang.org/x/tools v0.24.0
+	gotest.tools/v3 v3.5.1
 )
 
-go 1.13
+require (
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
+	golang.org/x/mod v0.20.0 // indirect
+	golang.org/x/text v0.17.0 // indirect
+)
