@@ -26,7 +26,7 @@ type writeSyncer interface {
 	Sync() error
 }
 
-// nolint:errcheck
+//nolint:errcheck
 func (h *eventHandler) Err(text string) error {
 	h.err.WriteString(text)
 	h.err.WriteRune('\n')
@@ -161,6 +161,7 @@ func postRunHook(opts *options, execution *testjson.Execution) error {
 	if len(command) == 0 {
 		return nil
 	}
+	log.Debugf("exec: %s", command)
 
 	cmd := exec.Command(command[0], command[1:]...)
 	cmd.Stdout = opts.stdout
@@ -170,7 +171,7 @@ func postRunHook(opts *options, execution *testjson.Execution) error {
 		"GOTESTSUM_JSONFILE="+opts.jsonFile,
 		"GOTESTSUM_JSONFILE_TIMING_EVENTS="+opts.jsonFileTimingEvents,
 		"GOTESTSUM_JUNITFILE="+opts.junitFile,
-		fmt.Sprintf("GOTESTSUM_ELAPSED=%.3f", execution.Elapsed().Seconds()),
+		fmt.Sprintf("GOTESTSUM_ELAPSED=%.3fs", execution.Elapsed().Seconds()),
 		fmt.Sprintf("TESTS_TOTAL=%d", execution.Total()),
 		fmt.Sprintf("TESTS_FAILED=%d", len(execution.Failed())),
 		fmt.Sprintf("TESTS_SKIPPED=%d", len(execution.Skipped())),
