@@ -16,7 +16,6 @@ import (
 
 func dotsFormatV1(out io.Writer) EventFormatter {
 	buf := bufio.NewWriter(out)
-	// nolint:errcheck
 	return eventFormatterFunc(func(event TestEvent, exec *Execution) error {
 		pkg := exec.Package(event.Package)
 		switch {
@@ -118,7 +117,7 @@ func (d *dotFormatter) Format(event TestEvent, exec *Execution) error {
 		pkgname := RelativePackagePath(pkg) + " "
 		prefix := fmtDotElapsed(exec.Package(pkg))
 		line.checkWidth(len(prefix+pkgname), d.termWidth)
-		fmt.Fprintf(d.writer, prefix+pkgname+line.builder.String()+"\n")
+		fmt.Fprint(d.writer, prefix+pkgname+line.builder.String()+"\n")
 	}
 	PrintSummary(d.writer, exec, SummarizeNone)
 	return d.writer.Flush()
