@@ -4,8 +4,11 @@ import (
 	"bufio"
 	"fmt"
 	"io"
+	"os"
+	"sort"
 	"strings"
 
+	"github.com/bitfield/gotestdox"
 	"github.com/fatih/color"
 )
 
@@ -65,7 +68,7 @@ func standardQuietFormat(out io.Writer) EventFormatter {
 // go test -json
 func standardJSONFormat(out io.Writer) EventFormatter {
 	buf := bufio.NewWriter(out)
-	// nolint:errcheck // errors are returned by Flush
+	//nolint:errcheck // errors are returned by Flush
 	return eventFormatterFunc(func(event TestEvent, _ *Execution) error {
 		buf.Write(event.raw)
 		buf.WriteRune('\n')
@@ -73,18 +76,74 @@ func standardJSONFormat(out io.Writer) EventFormatter {
 	})
 }
 
+func testNameFormatTestEvent(out io.Writer, event TestEvent) {
+	pkgPath := RelativePackagePath(event.Package)
+
+	fmt.Fprintf(out, "%s %s%s (%.2fs)\n",
+		colorEvent(event)(strings.ToUpper(string(event.Action))),
+		joinPkgToTestName(pkgPath, event.Test),
+		formatRunID(event.RunID),
+		event.Elapsed)
+}
+
+func testDoxFormat(out io.Writer, opts FormatOptions) EventFormatter {
+	buf := bufio.NewWriter(out)
+	type Result struct {
+		Event    TestEvent
+		Sentence string
+	}
+
+	getIcon := getIconFunc(opts)
+	results := map[string][]Result{}
+	return eventFormatterFunc(func(event TestEvent, _ *Execution) error {
+		switch {
+		case event.PackageEvent():
+			if !event.Action.IsTerminal() {
+				return nil
+			}
+			if opts.HideEmptyPackages && len(results[event.Package]) == 0 {
+				return nil
+			}
+			fmt.Fprintf(buf, "%s:\n", event.Package)
+			tests := results[event.Package]
+			sort.Slice(tests, func(i, j int) bool {
+				return tests[i].Sentence < tests[j].Sentence
+			})
+			for _, r := range tests {
+				fmt.Fprintf(buf, " %s %s (%.2fs)\n",
+					getIcon(r.Event.Action),
+					r.Sentence,
+					r.Event.Elapsed)
+			}
+			fmt.Fprintln(buf)
+			return buf.Flush()
+		case event.Action.IsTerminal():
+			// Fuzz test cases tend not to have interesting names,
+			// so only report these if they're failures
+			if isFuzzCase(event) {
+				return nil
+			}
+			results[event.Package] = append(results[event.Package], Result{
+				Event:    event,
+				Sentence: gotestdox.Prettify(event.Test),
+			})
+		}
+		return nil
+	})
+}
+
+func isFuzzCase(event TestEvent) bool {
+	return strings.HasPrefix(event.Test, "Fuzz") &&
+		event.Action == ActionPass &&
+		TestName(event.Test).IsSubTest()
+}
+
 func testNameFormat(out io.Writer) EventFormatter {
 	buf := bufio.NewWriter(out)
-	// nolint:errcheck
+	//nolint:errcheck
 	return eventFormatterFunc(func(event TestEvent, exec *Execution) error {
 		formatTest := func() error {
-			pkgPath := RelativePackagePath(event.Package)
-
-			fmt.Fprintf(buf, "%s %s%s %s\n",
-				colorEvent(event)(strings.ToUpper(string(event.Action))),
-				joinPkgToTestName(pkgPath, event.Test),
-				formatRunID(event.RunID),
-				event.ElapsedFormatted())
+			testNameFormatTestEvent(buf, event)
 			return buf.Flush()
 		}
 
@@ -101,6 +160,7 @@ func testNameFormat(out io.Writer) EventFormatter {
 			result := colorEvent(event)(strings.ToUpper(string(event.Action)))
 			pkg := exec.Package(event.Package)
 			if event.Action == ActionSkip || (event.Action == ActionPass && pkg.Total == 0) {
+				event.Action = ActionSkip // always color these as skip actions
 				result = colorEvent(event)("EMPTY")
 			}
 
@@ -116,7 +176,7 @@ func testNameFormat(out io.Writer) EventFormatter {
 			pkg.WriteOutputTo(buf, tc.ID)
 			return formatTest()
 
-		case event.Action == ActionPass:
+		case event.Action == ActionPass || event.Action == ActionSkip:
 			return formatTest()
 		}
 		return nil
@@ -182,40 +242,109 @@ func pkgNameFormat(out io.Writer, opts FormatOptions) eventFormatterFunc {
 	}
 }
 
+type icons struct {
+	pass  string
+	fail  string
+	skip  string
+	color bool
+}
+
+func (i icons) forAction(action Action) string {
+	if i.color {
+		switch action {
+		case ActionPass:
+			return color.GreenString(i.pass)
+		case ActionSkip:
+			return color.YellowString(i.skip)
+		case ActionFail:
+			return color.RedString(i.fail)
+		default:
+			return " "
+		}
+	} else {
+		switch action {
+		case ActionPass:
+			return i.pass
+		case ActionSkip:
+			return i.skip
+		case ActionFail:
+			return i.fail
+		default:
+			return " "
+		}
+	}
+}
+
+func getIconFunc(opts FormatOptions) func(Action) string {
+	switch {
+	case opts.UseHiVisibilityIcons || opts.Icons == "hivis":
+		return icons{
+			pass:  "✅", // WHITE HEAVY CHECK MARK
+			skip:  "➖", // HEAVY MINUS SIGN
+			fail:  "❌", // CROSS MARK
+			color: false,
+		}.forAction
+	case opts.Icons == "text":
+		return icons{
+			pass:  "PASS",
+			skip:  "SKIP",
+			fail:  "FAIL",
+			color: true,
+		}.forAction
+	case opts.Icons == "codicons":
+		return icons{
+			pass:  "\ueba4", // cod-pass
+			skip:  "\ueabd", // cod-circle_slash
+			fail:  "\uea87", // cod-error
+			color: true,
+		}.forAction
+	case opts.Icons == "octicons":
+		return icons{
+			pass:  "\uf49e", // oct-check_circle
+			skip:  "\uf517", // oct-skip
+			fail:  "\uf52f", // oct-x_circle
+			color: true,
+		}.forAction
+	case opts.Icons == "emoticons":
+		return icons{
+			pass:  "\U000f01f5", // md-emoticon_happy_outline
+			skip:  "\U000f01f6", // md-emoticon_neutral_outline
+			fail:  "\U000f01f8", // md-emoticon_sad_outline
+			color: true,
+		}.forAction
+	default:
+		return icons{
+			pass:  "✓", // CHECK MARK
+			skip:  "∅", // EMPTY SET
+			fail:  "✖", // HEAVY MULTIPLICATION X
+			color: true,
+		}.forAction
+	}
+}
+
 func shortFormatPackageEvent(opts FormatOptions, event TestEvent, exec *Execution) string {
 	pkg := exec.Package(event.Package)
 
-	var iconSkipped, iconSuccess, iconFailure string
-	if opts.UseHiVisibilityIcons {
-		iconSkipped = "➖"
-		iconSuccess = "✅"
-		iconFailure = "❌"
-	} else {
-		iconSkipped = "∅"
-		iconSuccess = "✓"
-		iconFailure = "✖"
-	}
-
+	getIcon := getIconFunc(opts)
 	fmtEvent := func(action string) string {
 		return action + "  " + packageLine(event, exec.Package(event.Package))
 	}
-	withColor := colorEvent(event)
 	switch event.Action {
 	case ActionSkip:
 		if opts.HideEmptyPackages {
 			return ""
 		}
-		return fmtEvent(withColor(iconSkipped))
+		return fmtEvent(getIcon(event.Action))
 	case ActionPass:
 		if pkg.Total == 0 {
 			if opts.HideEmptyPackages {
 				return ""
 			}
-			return fmtEvent(withColor(iconSkipped))
+			return fmtEvent(getIcon(ActionSkip))
 		}
-		return fmtEvent(withColor(iconSuccess))
+		return fmtEvent(getIcon(event.Action))
 	case ActionFail:
-		return fmtEvent(withColor(iconFailure))
+		return fmtEvent(getIcon(event.Action))
 	}
 	return ""
 }
@@ -250,12 +379,12 @@ func pkgNameWithFailuresFormat(out io.Writer, opts FormatOptions) eventFormatter
 			if event.Action == ActionFail {
 				pkg := exec.Package(event.Package)
 				tc := pkg.LastFailedByName(event.Test)
-				pkg.WriteOutputTo(buf, tc.ID) // nolint:errcheck
+				pkg.WriteOutputTo(buf, tc.ID) //nolint:errcheck
 				return buf.Flush()
 			}
 			return nil
 		}
-		buf.WriteString(shortFormatPackageEvent(opts, event, exec)) // nolint:errcheck
+		buf.WriteString(shortFormatPackageEvent(opts, event, exec))
 		return buf.Flush()
 	}
 }
@@ -286,7 +415,8 @@ func (e eventFormatterFunc) Format(event TestEvent, output *Execution) error {
 
 type FormatOptions struct {
 	HideEmptyPackages    bool
-	UseHiVisibilityIcons bool
+	UseHiVisibilityIcons bool // Deprecated
+	Icons                string
 }
 
 // NewEventFormatter returns a formatter for printing events.
@@ -306,13 +436,80 @@ func NewEventFormatter(out io.Writer, format string, formatOpts FormatOptions) E
 		return dotsFormatV1(out)
 	case "dots-v2":
 		return newDotFormatter(out, formatOpts)
+	case "gotestdox", "testdox":
+		return testDoxFormat(out, formatOpts)
 	case "testname", "short-verbose":
+		if os.Getenv("GITHUB_ACTIONS") == "true" {
+			return githubActionsFormat(out)
+		}
 		return testNameFormat(out)
 	case "pkgname", "short":
 		return pkgNameFormat(out, formatOpts)
 	case "pkgname-and-test-fails", "short-with-failures":
 		return pkgNameWithFailuresFormat(out, formatOpts)
+	case "github-actions", "github-action":
+		return githubActionsFormat(out)
 	default:
 		return nil
 	}
+}
+
+func githubActionsFormat(out io.Writer) EventFormatter {
+	buf := bufio.NewWriter(out)
+
+	type name struct {
+		Package string
+		Test    string
+	}
+	output := map[name][]string{}
+
+	return eventFormatterFunc(func(event TestEvent, exec *Execution) error {
+		key := name{Package: event.Package, Test: event.Test}
+
+		// test case output
+		if event.Test != "" && event.Action == ActionOutput {
+			if !isFramingLine(event.Output, event.Test) {
+				output[key] = append(output[key], event.Output)
+			}
+			return nil
+		}
+
+		// test case end event
+		if event.Test != "" && event.Action.IsTerminal() {
+			if len(output[key]) > 0 {
+				buf.WriteString("::group::")
+			} else {
+				buf.WriteString("  ")
+			}
+			testNameFormatTestEvent(buf, event)
+
+			for _, item := range output[key] {
+				buf.WriteString(item)
+			}
+			if len(output[key]) > 0 {
+				buf.WriteString("\n::endgroup::\n")
+			}
+			delete(output, key)
+			return buf.Flush()
+		}
+
+		// package event
+		if !event.Action.IsTerminal() {
+			return nil
+		}
+
+		result := colorEvent(event)(strings.ToUpper(string(event.Action)))
+		pkg := exec.Package(event.Package)
+		if event.Action == ActionSkip || (event.Action == ActionPass && pkg.Total == 0) {
+			event.Action = ActionSkip // always color these as skip actions
+			result = colorEvent(event)("EMPTY")
+		}
+
+		buf.WriteString("  ")
+		buf.WriteString(result)
+		buf.WriteString(" Package ")
+		buf.WriteString(packageLine(event, exec.Package(event.Package)))
+		buf.WriteString("\n")
+		return buf.Flush()
+	})
 }
