@@ -87,6 +87,13 @@ func TestFormats_DefaultGoTestJson(t *testing.T) {
 
 	testCases := []testCase{
 		{
+			name: "testdox",
+			format: func(out io.Writer) EventFormatter {
+				return testDoxFormat(out, FormatOptions{})
+			},
+			expectedOut: "format/testdox.out",
+		},
+		{
 			name:        "testname",
 			format:      testNameFormat,
 			expectedOut: "format/testname.out",
@@ -106,9 +113,37 @@ func TestFormats_DefaultGoTestJson(t *testing.T) {
 		{
 			name: "pkgname with hivis",
 			format: func(out io.Writer) EventFormatter {
-				return pkgNameFormat(out, FormatOptions{UseHiVisibilityIcons: true})
+				return pkgNameFormat(out, FormatOptions{Icons: "hivis"})
 			},
 			expectedOut: "format/pkgname-hivis.out",
+		},
+		{
+			name: "pkgname with text",
+			format: func(out io.Writer) EventFormatter {
+				return pkgNameFormat(out, FormatOptions{Icons: "text"})
+			},
+			expectedOut: "format/pkgname-text.out",
+		},
+		{
+			name: "pkgname with codicons",
+			format: func(out io.Writer) EventFormatter {
+				return pkgNameFormat(out, FormatOptions{Icons: "codicons"})
+			},
+			expectedOut: "format/pkgname-codicons.out",
+		},
+		{
+			name: "pkgname with octicons",
+			format: func(out io.Writer) EventFormatter {
+				return pkgNameFormat(out, FormatOptions{Icons: "octicons"})
+			},
+			expectedOut: "format/pkgname-octicons.out",
+		},
+		{
+			name: "pkgname with emoticons",
+			format: func(out io.Writer) EventFormatter {
+				return pkgNameFormat(out, FormatOptions{Icons: "emoticons"})
+			},
+			expectedOut: "format/pkgname-emoticons.out",
 		},
 		{
 			name: "pkgname with hide-empty",
@@ -131,6 +166,11 @@ func TestFormats_DefaultGoTestJson(t *testing.T) {
 			name:        "standard-json",
 			format:      standardJSONFormat,
 			expectedOut: "input/go-test-json.out",
+		},
+		{
+			name:        "github-actions",
+			format:      githubActionsFormat,
+			expectedOut: "format/github-actions.out",
 		},
 	}
 
@@ -171,6 +211,13 @@ func TestFormats_Coverage(t *testing.T) {
 	}
 
 	testCases := []testCase{
+		{
+			name: "testdox",
+			format: func(out io.Writer) EventFormatter {
+				return testDoxFormat(out, FormatOptions{})
+			},
+			expectedOut: "format/testdox-coverage.out",
+		},
 		{
 			name:        "testname",
 			format:      testNameFormat,
@@ -239,6 +286,13 @@ func TestFormats_Shuffle(t *testing.T) {
 	}
 
 	testCases := []testCase{
+		{
+			name: "testdox",
+			format: func(out io.Writer) EventFormatter {
+				return testDoxFormat(out, FormatOptions{})
+			},
+			expectedOut: "format/testdox-shuffle.out",
+		},
 		{
 			name:        "testname",
 			format:      testNameFormat,
