@@ -14,7 +14,6 @@ import (
 // Summary enumerates the sections which can be printed by PrintSummary
 type Summary int
 
-// nolint: golint
 const (
 	SummarizeNone    Summary = 0
 	SummarizeSkipped Summary = (1 << iota) / 2
@@ -178,7 +177,7 @@ func writeTestCaseSummary(out io.Writer, execution executionSummary, conf testCa
 			formatRunID(tc.RunID),
 			FormatDurationAsSeconds(tc.Elapsed, 2))
 		for _, line := range execution.OutputLines(tc) {
-			if isFramingLine(line) || conf.filter(tc.Test.Name(), line) {
+			if isFramingLine(line, tc.Test.Name()) {
 				continue
 			}
 			fmt.Fprint(out, line)
@@ -192,7 +191,6 @@ func writeTestCaseSummary(out io.Writer, execution executionSummary, conf testCa
 type testCaseFormatConfig struct {
 	header string
 	prefix string
-	filter func(testName string, line string) bool
 	getter func(executionSummary) []TestCase
 }
 
@@ -201,9 +199,6 @@ func formatFailed() testCaseFormatConfig {
 	return testCaseFormatConfig{
 		header: withColor("Failed"),
 		prefix: withColor("FAIL"),
-		filter: func(testName string, line string) bool {
-			return strings.HasPrefix(line, "--- FAIL: "+testName+" ")
-		},
 		getter: func(execution executionSummary) []TestCase {
 			return execution.Failed()
 		},
@@ -215,17 +210,17 @@ func formatSkipped() testCaseFormatConfig {
 	return testCaseFormatConfig{
 		header: withColor("Skipped"),
 		prefix: withColor("SKIP"),
-		filter: func(testName string, line string) bool {
-			return strings.HasPrefix(line, "--- SKIP: "+testName+" ")
-		},
 		getter: func(execution executionSummary) []TestCase {
 			return execution.Skipped()
 		},
 	}
 }
 
-func isFramingLine(line string) bool {
+func isFramingLine(line string, testName string) bool {
 	return strings.HasPrefix(line, "=== RUN   Test") ||
 		strings.HasPrefix(line, "=== PAUSE Test") ||
-		strings.HasPrefix(line, "=== CONT  Test")
+		strings.HasPrefix(line, "=== CONT  Test") ||
+		strings.HasPrefix(line, "--- FAIL: "+testName+" ") ||
+		strings.HasPrefix(line, "--- SKIP: "+testName+" ") ||
+		strings.HasPrefix(line, "--- PASS: "+testName+" ")
 }
